+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - Add a first-time setup script.
 - Add support for NVIDIA graphics drivers.
 - Add support for Intel OpenGL graphics drivers on SteamOS.
+- Add support for changing the locale as part of the first-time setup.
 
 ## 2.2.0